feat: Implement FamilyCard component and integrate it into Family management screen
